// lib/gen/schema/user.ts
/**
 * Generated by Orval v7.0.0
 * Date: 2025-12-19T18:00:52.062Z
 * Title: FastAPI Server
 * Version: 1.0.0
 */
import type { Role } from "./role";

/**
 * Public user model without sensitive data
 */
export interface User {
  id: string;
  first_name: string;
  last_name: string;
  email: string;
  username: string;
  roles?: Role[];
  created_at: string;
  updated_at: string;
  is_active?: boolean;
}